Moved mock_font into its own module
